ready for test and norminette

+++ philo/sources/check_args/check_args.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   check_args.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: fberthou <fberthou@student.42.fr>          +#+  +:+       +#+        */
+/*   By: florian <florian@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/23 13:46:59 by fberthou          #+#    #+#             */
-/*   Updated: 2024/07/17 11:09:38 by fberthou         ###   ########.fr       */
+/*   Updated: 2024/07/19 17:33:07 by florian          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,9 +36,9 @@ static bool ft_isdigit(char *arg)
 
 int	check_args(int argc, char **argv, int *tab_arg)
 {
-	size_t		i_argv;
-	short		  i_tab_arg;
-	long int	tmp;
+	int         i_argv;
+	int         i_tab_arg;
+	long int    tmp;
 
 	i_argv = 1;
 	i_tab_arg = 0;
